+++ R/Problem.R
@@ -64,7 +64,7 @@
 #' tmp$algorithms
 #' getJobPars(reg = tmp)
 addProblem = function(name, data = NULL, fun = NULL, seed = NULL, cache = FALSE, reg = getDefaultRegistry()) {
-  assertExperimentRegistry(reg, writeable = TRUE)
+  assertRegistry(reg, class = "ExperimentRegistry", writeable = TRUE)
   assertString(name, min.chars = 1L)
   if (!stri_detect_regex(name, "^[[:alnum:]_.-]+$"))
     stopf("Illegal characters in problem name: %s", name)
@@ -96,7 +96,7 @@ addProblem = function(name, data = NULL, fun = NULL, seed = NULL, cache = FALSE,
 #' @export
 #' @rdname addProblem
 removeProblems = function(name, reg = getDefaultRegistry()) {
-  assertExperimentRegistry(reg, writeable = TRUE, running.ok = FALSE)
+  assertRegistry(reg, class = "ExperimentRegistry", writeable = TRUE, running.ok = FALSE)
   assertCharacter(name, any.missing = FALSE)
   assertSubset(name, reg$problems)