+++ R/mergeRegistries.R
@@ -32,8 +32,8 @@
 #
 # new = mergeRegistries(source, target)
 mergeRegistries = function(source, target = getDefaultRegistry()) {
-  assertRegistry(source, writeable = TRUE, running.ok = FALSE)
-  assertRegistry(target, writeable = TRUE, running.ok = FALSE)
+  assertRegistry(source, writeable = TRUE, sync = TRUE, running.ok = FALSE)
+  assertRegistry(target, writeable = TRUE, sync = TRUE, running.ok = FALSE)
   if (normalizePath(source$file.dir, winslash = "/") == normalizePath(target$file.dir, winslash = "/"))
     stop("You must provide two different registries (using different file directories")
   hash = function(x) unlist(.mapply(function(...) digest(list(...)), x[, !"def.id"], list()))